pyproject.toml
[project]
name = "benchmark-annotation"
version = "0.1.0"
description = "DOCX markup placement evaluator CLI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-docx>=1.0.0",
  "pdf2image>=1.16.3",
  "Pillow>=10.0.0",
]

[project.scripts]
docx-markup-eval = "docx_markup_eval.cli:main"

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]

[tool.uv]
package = true
dev-dependencies = ["pytest>=8.0.0"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
